+++ src/EventGrid/EventGrid/AzureEventGridCmdletBase.cs
@@ -80,12 +80,12 @@ namespace Microsoft.Azure.Commands.EventGrid
         protected const string DomainTopicNameParameterSet = "DomainTopicNameParameterSet";
 
         protected const string VerifiedPartnerNameParameterSet = "VerifiedPartnerNameParameterSet";
-        protected const string PartnerConfigurationListBySubscriptionParameterSet = "PartnerConfigurationNameParameterSet";
+        protected const string PartnerConfigurationListBySubscriptionParameterSet = "PartnerConfigurationListBySubscriptionParameterSet";
         protected const string PartnerRegistrationNameParameterSet = "PartnerRegistrationNameParameterSet";
-        protected const string PartnerRegistrationListBySubscriptionParameterSet = "PartnerRegistrationNameParameterSet";
+        protected const string PartnerRegistrationListBySubscriptionParameterSet = "PartnerRegistrationListBySubscriptionParameterSet";
         protected const string PartnerDestinationNameParameterSet = "PartnerDestinationNameParameterSet";
         protected const string PartnerNamespaceNameParameterSet = "PartnerNamespaceNameParameterSet";
-        protected const string PartnerNamespaceListBySubscriptionParameterSet = "PartnerConfigurationNameParameterSet";
+        protected const string PartnerNamespaceListBySubscriptionParameterSet = "PartnerNamespaceListBySubscriptionParameterSet";
         protected const string ChannelNameParameterSet = "ChannelNameParameterSet";
         protected const string ChannelListByPartnerNamespaceParameterSet = "ChannelListByPartnerNamespaceParameterSet";
         protected const string PartnerTopicNameParameterSet = "PartnerTopicNameParameterSet";